Fetching more parameters in Importer; Now all Solids must be created with a position; Solids are no longer squares but circles; Changed speed, dtime, angle to double; Removed some IDEA warnings;

+++ simulator/src/com/md/mechevo/game/action/Action.java
@@ -111,11 +111,11 @@ public abstract class Action implements EventObservable {
 
 	/**
 	 * Execute the action.
-	 * 
+	 *
 	 * @param state Current State of the game
 	 * @param dtime Duration of the round
 	 */
-	public abstract void update(State state, float dtime);
+	public abstract void update(State state, double dtime);
 
 
 	/**
